requirements.txt
filelock==3.16.1
fsspec==2024.12.0
Jinja2==3.1.5
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==1.26.4
pygame==2.6.1
PyYAML==6.0.2
sympy==1.13.3
torch==2.2.2
typing_extensions==4.12.2
